Move o handler para o lugar correto.
