+++ src/Board/setup.cpp
@@ -1,43 +1,5 @@
 #include "FTRLab/Board.hpp"
 
-void Board::CommunicationHandler()
-{
-  TickType_t xLastWakeTime = xTaskGetTickCount();
-  BaseType_t xWasDelayed;
-
-  Serial.println("CommunicationTask running on core " + String(xPortGetCoreID()));
-
-  while (true)
-  {
-    if (WiFi.status() == WL_CONNECTED)
-    {
-      digitalWrite(this->pins.networkLed, HIGH);
-
-      if (!this->client.connected())
-      {
-        this->client = this->server.available(); // Disponibiliza o servidor para o cliente se conectar.
-        delay(100);
-      }
-    }
-    else
-    {
-      digitalWrite(this->pins.networkLed, LOW);
-
-      this->connectToNetwork();
-      this->printNetworkInfo();
-      this->setupNTP();
-      this->server.begin();
-      this->setupMdns();
-    }
-
-    while (WiFi.status() == WL_CONNECTED && this->client.connected())
-    {
-      xWasDelayed = xTaskDelayUntil(&xLastWakeTime, pdMS_TO_TICKS(this->targetSendingPeriod));
-      this->sendAllMeasurements();
-    }
-  }
-}
-
 void CommunicationHandlerWrapper(void *pvParameters)
 {
   board.CommunicationHandler();
@@ -53,7 +15,7 @@ void Board::setup()
 
   xTaskCreatePinnedToCore(
       CommunicationHandlerWrapper, /* Task function. */
-      "CommunicationTask",         /* name of task. */
+      "Communication",             /* name of task. */
       10000,                       /* Stack size of task */
       NULL,                        /* parameter of the task */
       tskIDLE_PRIORITY,            /* priority of the task */
